requirements.txt
requests==2.31.0
pandas==2.1.1
numpy==1.26.0
matplotlib==3.8.0
seaborn==0.13.0
sqlalchemy==2.0.23
biopython==1.81
pytest==7.4.3
python-dotenv==1.0.0 